Refactor overlay backend into cooperative domain mixins
